refactored newNote data struct to hold float64 for nostr event kinds

+++ data/newNote.go
@@ -1,8 +1,8 @@
 package data
 
 type NewNote struct {
-	PrivHexKey string `json:"privHexKey"`
-	PubHexKey  string `json:"pubHexKey"`
-	Kind       int    `json:"kind"`
-	Content    string `json:"content"`
+	PrivHexKey string  `json:"privHexKey"`
+	PubHexKey  string  `json:"pubHexKey"`
+	Kind       float64 `json:"kind"`
+	Content    string  `json:"content"`
 }
